LayoutTests/imported/mozilla/css-transitions/test_effect-target-expected.txt


PASS Returned CSS transitions have the correct Animation.target 
FAIL effect.target should return the same CSSPseudoElement object each time assert_equals: Got transitions running on ::after pseudo element expected 2 but got 0
FAIL effect.target from the script-generated animation should return the same CSSPseudoElement object as that from the CSS generated transition undefined is not an object (evaluating 'document.getAnimations()[0].effect')
